Hero component built, Layouts restructured
