+++ src/components/atoms/index.js
@@ -3,3 +3,4 @@ export { ScrollContainer } from './ScrollContainer/ScrollContainer'
 export { SectionTitle } from './SectionTitle/SectionTitle'
 export { Loader } from './Loader/Loader'
 export { PageWrapper } from './PageWrapper/PageWrapper'
+export { Hero } from './Hero/Hero'
