+++ src/index.html
@@ -7,6 +7,12 @@
         <meta name="theme-color" content="#000000" />
         <meta name="description" content="Video game content and news." />
         <link rel="stylesheet" href="./App.css" type="text/css" />
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+        <link
+            href="https://fonts.googleapis.com/css2?family=Inter&family=Open+Sans&family=Roboto&display=swap"
+            rel="stylesheet"
+        />
         <title>Video Games Central</title>
     </head>
     <body>
