+++ src/routes/Layouts/MainLayout/MainLayout.module.scss
@@ -1,0 +1,5 @@
+@use 'src/styles/placeholders';
+
+.layout {
+    @extend %flex-row;
+}
