+++ src/routes/Layouts/__tests__/MainLayout.test.js
@@ -1,8 +1,8 @@
 import { waitFor } from '@testing-library/react'
 import axios from 'axios'
-import { FAKE_NEWS } from '../../data/dummy-data'
-import { renderWithClient } from '../../utils/testUtils'
-import { Layout } from '../Layout/Layout'
+import { FAKE_NEWS } from '../../../data/dummy-data'
+import { renderWithClient } from '../../../utils/testUtils'
+import { MainLayout } from '../MainLayout/MainLayout'
 
 beforeEach(() => {
     axios.get.mockResolvedValue({ data: FAKE_NEWS })
@@ -10,7 +10,7 @@ beforeEach(() => {
 
 describe('<Layout />', () => {
     it('renders', async () => {
-        const { getByText, getByTestId } = renderWithClient(<Layout />)
+        const { getByText, getByTestId } = renderWithClient(<MainLayout />)
 
         await waitFor(() => {
             // Ensures navbar renders
