+++ src/routes/Layouts/index.js
@@ -1,0 +1,2 @@
+export { MainLayout } from './MainLayout/MainLayout'
+export { HomeLayout } from './HomeLayout/HomeLayout'
