+++ src/styles/placeholders.scss
@@ -38,6 +38,10 @@
     font-family: 'Roboto', sans-serif;
 }
 
+%font-inter {
+    font-family: 'Inter', sans-serif;
+}
+
 %text-ellipsis {
     overflow: hidden;
     text-overflow: ellipsis;
