+++ src/styles/vars.scss
@@ -3,6 +3,7 @@ $navbar-width: 300px;
 $page-width: calc(100% - $navbar-width);
 $page-padding: 50px;
 $ratingBoxSize: 50px;
+$hero-padding: 3rem;
 
 // Colors
 // Got these color names from https://www.color-name.com/hex.
